requirements.txt
aniso8601==7.0.0
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudinary==1.20.0
cryptography==2.9.2
Flask==1.1.2
Flask-GraphQL==2.0.1
flask-mongoengine==0.9.5
Flask-WTF==0.14.3
graphene==2.1.8
graphql-core==2.3.1
graphql-relay==2.0.1
graphql-server-core==1.2.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
midtransclient==1.1.1
mongoengine==0.20.0
promise==2.3
pycparser==2.20
PyJWT==1.7.1
pymongo==3.10.1
python-dotenv==0.13.0
requests==2.23.0
Rx==1.6.1
six==1.14.0
urllib3==1.25.9
Werkzeug==1.0.1
WTForms==2.3.1
